Fixes to keyboard shortcuts and solo route
